Move accumulators to separate file
